move Value to pkg/ast
Creates a new `pkg/ast` package that will hold the Abstract Syntax Tree
(AST) nodes. I previously wanted to put these in `pkg/grammar` but since
these files import `pkg/types` and `pkg/types` imports `pkg/grammar`, we
ran into a cyclic import failure.

The first AST node to move into the new `pkg/ast` package is the simple
`Value` struct, which is a placeholder for a literal value in a SQL
statement.
