+++ function.go
@@ -6,6 +6,7 @@
 package sqlb
 
 import (
+	"github.com/jaypipes/sqlb/pkg/ast"
 	"github.com/jaypipes/sqlb/pkg/grammar"
 	"github.com/jaypipes/sqlb/pkg/types"
 )
@@ -243,7 +244,7 @@ func ConcatWs(sep string, projs ...types.Projection) *sqlFunc {
 	subjects := &List{elements: els}
 	return &sqlFunc{
 		ScanInfo: grammar.FunctionScanTable(grammar.FUNC_CONCAT_WS),
-		elements: []types.Element{&Value{val: sep}, subjects},
+		elements: []types.Element{ast.NewValue(nil, sep), subjects},
 		// TODO(jaypipes): Clearly we need to support >1 selection...
 		sel: projs[0].From(),
 	}
