+++ select_statement_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/jaypipes/sqlb/pkg/ast"
 	"github.com/jaypipes/sqlb/pkg/scanner"
 	"github.com/jaypipes/sqlb/pkg/testutil"
 	"github.com/jaypipes/sqlb/pkg/types"
@@ -39,7 +40,7 @@ func TestSelectClause(t *testing.T) {
 		{
 			name: "A literal value",
 			s: &SelectStatement{
-				projs: []types.Projection{&Value{val: 1}},
+				projs: []types.Projection{ast.NewValue(nil, 1)},
 			},
 			qs:    "SELECT ?",
 			qargs: []interface{}{1},
@@ -48,7 +49,7 @@ func TestSelectClause(t *testing.T) {
 			name: "A literal value aliased",
 			s: &SelectStatement{
 				projs: []types.Projection{
-					&Value{alias: "foo", val: 1},
+					ast.NewValue(nil, 1).As("foo"),
 				},
 			},
 			qs:    "SELECT ? AS foo",
@@ -58,8 +59,8 @@ func TestSelectClause(t *testing.T) {
 			name: "Two literal values",
 			s: &SelectStatement{
 				projs: []types.Projection{
-					&Value{val: 1},
-					&Value{val: 1},
+					ast.NewValue(nil, 1),
+					ast.NewValue(nil, 1),
 				},
 			},
 			qs:    "SELECT ?, ?",
